Intento corregir lo de consensuado

+++ agregador/src/main/java/ar/edu/utn/frba/ddsi/agregador/models/entities/hecho/Hecho.java
@@ -98,7 +98,7 @@ public abstract class Hecho {
 
 
     public boolean consensuado(Double mencionesNecesarias) {
-        return this.cantidadMenciones >= mencionesNecesarias;
+        return this.cantidadMenciones >= mencionesNecesarias.intValue();
     }
 }
